Truncate any log line longer than 99KB before forwarding to GentlemanJerry.
A customer was writing 20MB lines to their log files. This caused JoeCool
instances to use > 20MB of memory reading those lines in and caused
GentlemanJerry to get into a bad state, constantly restarting and unable to
forward logs to Papertrail, since Papertrail doesn't accept messages > 100 KB.

This patch uses a yet-to-be-merged patch to logstash-forwarder
(https://github.com/aaw/logstash-forwarder/commit/141d0c5) to enforce a max
line length of 99KB in JoeCool before messages get truncated. The truncation
happens in the logstash-forwarder so not only do large messages not get sent
to logstash, they are not even held in memory in the forwarder.

The magic number 99KB was chosen because syslog messages contain metadata
like source, severity, etc. This metadata should never be more than 1KB, so
a 99KB bound is large enough to be reasonable for any logs that we want to
forward and it (hopefully) guarantees that Papertrail should always accept
the logs we forward.