Move to alpine, fix bug where files are ignored after 1 day of inactivity.
Adding a "dead time" value of 720 hours (30 * 24 hours), which should resolve
issue #5. Dead time has only been recently exposed as a config parameter in 1.5.

Moving to alpine as a base image. It was a pretty easy transition and cuts the
image size down from ~500 MB to ~25 MB. We could cut more if we removed ruby
(we're only using erb right now to generate the config file, so we'd just need
to replace that with any simpler templating system).